Created bullet + bullet effect logic + touch tap control system
The bullet effect system is based on the idea that each effect can modify/override any property on the gun or the bullets in the gun. Example: SplitBulletEffect doubles all the bullets shot every time the gun is fired + creates individual spawn points for them. Each of these bullets are also modified by the LinePathBulletEffect that makes each of the bullets fly up in a straight line.

Control now depends on user tapping left and right sides of the screen